+++ jest/mockInstabug.js
@@ -45,7 +45,6 @@ jest.mock('NativeModules', () => {
             setPreSendingHandler: jest.fn(),
             callPrivateApi: jest.fn(),
             addListener: jest.fn(),
-            submitReport: jest.fn(),
             getReport: jest.fn(),
             sendHandledJSCrash: jest.fn(),
             sendJSCrash: jest.fn(),