+++ ios/Instabug.framework/Headers/IBGTypes.h
@@ -5,7 +5,7 @@
  
  Copyright:  (c) 2013-2018 by Instabug, Inc., all rights reserved.
  
- Version:    9.1.6
+ Version:    9.1.7
  */
 
 #import <UIKit/UIKit.h>
@@ -279,6 +279,7 @@ typedef NS_ENUM(NSInteger, IBGDismissType) {
  */
 typedef NS_ENUM(NSInteger, IBGLocale) {
     IBGLocaleArabic,
+    IBGLocaleAzerbaijani,
     IBGLocaleChineseSimplified,
     IBGLocaleChineseTaiwan,
     IBGLocaleChineseTraditional,