+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* Fixes an issue with `onReportSubmitHandler` on iOS
+
 ## v9.1.7 (2020-08-10)
 
 * Fixes missing typescript definitions